dct tests: run all possible sizes in one test
Modify fdct4x4_test.cc to support all size combinations. This does not
add any new tests and in fact fails a few. There were minimal changes
made to the tests so it's not entirely surprising that some of the
larger 12 bit transforms are failing.

In follow-up patches the tests in fdct8x8_test.cc and dct16x16_test.cc
will be evaluated and moved to dct_test.cc.

Change-Id: I72a23430f457d7fae8c91e706adc0e77c25abc8f