Update patch set 1
Patch Set 1:

That helps some but comments in the code and a more descriptive commit message would improve things a lot. I couldn't figure out what was going on with the old change based on the message and this style of appending seems pretty new, so calling it out in code comments and the commit message would be appreciated.

Patch-set: 1
Reviewer: Gerrit User 1117029 <1117029@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>