Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> it would probably be good to get the aligned allocations sorted out first, then do the moves and later add the other combinations to keep the size of the change down. moving to unaligned loads and stores isn't too risky, but it's a bit of unnecessary churn when it's only the test that's wrong

Can we review the rest in the meantime? That's a pretty minor part of this change. There doesn't seem to be much progress on the other patch, which means this is going to stall for a long time.

Patch-set: 4
Reviewer: Gerrit User 1117029 <1117029@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>