Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 6: Code-Review-1
> > 
> > Style Warnings:
> > 
> >    vp8/encoder/temporal_filter.c:23:  Include the directory when naming .h files  [build/include] [4]
> >    vp8/encoder/onyx_if.c:52:  Include the directory when naming .h files  [build/include] [4]
> 
> Should we care about these style warnings? Some old codes also use header files without path.

this is a case of being consistent with the existing code, which admittedly is not at all. I just used the most local example to be consistent with. Anything further we would probably want to do a global cleanup

Patch-set: 6
Reviewer: Gerrit User 1002219 <1002219@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>