vp8: Apply drop_overshoot for non-screen content case.
For 1 pass CBR mode:
Apply the logic for dropping (and re-adjusting rate control)
due to large overshoot to the case of non-screen content when
drop frame is allowed.

For the non-screen content casei: add additional condition that
rate correction factor is close to minimum state, and add
flag to constrain the frequency of the dropping.

This feature is avoids large frame sizes after big content change
following low content period.

No change to behavior for screen_content_mode = 2.

Change-Id: I797ab236cbbf3b15cad439e9a227fbebced632e6