Update patch set 2
Patch Set 2:

Specifically this patch relates to a problem with change I6923... doing forced re-codes and altering the Q range if there is a big rate miss for the current frame, because for speeds 3+ the current frame size is not available.

The observed slow down for speeds 0-2 is not I believe a problem with I6923.. per se but with the fact that the parameters in the unit tests give rise to what is in effect an invalid / broken arf pattern. As such I want to keep the two things separate.

Patch-set: 2
Reviewer: Gerrit User 1000899 <1000899@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>