quantize ssse3: copy implementation to intrinsics
Removes early exit (pmovmskb) because the test shows it was *slower*
when coeff < zbin, which should be the easy skip case.

New code is marginally faster due to removing that and the skip_block
situation.

Still does not pass tests. Does match the assembly, although saving the
sign before multiplying is dubious.

Change-Id: Ia163f18c755aba542d6e93f7bf7343184660df5a