Remove the speed condition on scene detection in 1 pass code.
Scene detection is used for VBR mode and for screen_content mode.

It was also enabled for CBR mode via the speed condition,
but currently the analysis in the scene detection is not used
in CRB mode (similar computtations are done locally at superblock level
when source_sad feature is enabled).

No change in behavior.

Change-Id: I59991d7ef2af320bea7af4b907596e057affa42f