Experiment to improve grain retention.
This experiment adjusts the rate distortion and ARF filter based on the
spatial variance of the source and reconstruction.

In regards to the RD loop the method favors modes where the reconstruction
variance is similar to the source variance. However it is currently only
applied to regions where the source variance is quite low.

For very low variance blocks it applies a further bias against intra coding and large
prediction block sizes (the later in particular limit the usefulness of the loop filter).

The final part of this change is to lower the strength of the ARF filter for blocks
where the source has very low spatial variance, to encourage some low amplitude
texture or noise to pass through the filter.

This change improves the retention of film grain and fine noise/texture in spatially
flat regions, but as expected causes a significant drop in PSNR on many clips.
Tis is to be expected because similar but misaligned noise or texture will give a lower
PSNR than a flat noise free reconstruction. However, it is worth noting that most clips
show a strong gain in FAST SSIM.

For this initial upload the new features are disabled and the previous variance adjustment
is also disabled.

Change-Id: I26a4e4edfa3dc5cacead82fa701fe7a9118ccd0a