Prevent double application of min rate in corpus vbr.
The initial allocation of bits in the two pass code to each frame
should be within the min max limits on the command line. However,
when forming an ARF group the cost of the ARF is shared by frames
in that group such that the residual bits for a frame could drop below
the min value. This change for corpus VBR prevents the minimum
being re-applied after the cost of the ARF has been deducted as this
will otherwise cause low rate sections to overshoot their target.

Change-Id: I715da7b130bf53ba8aa609532aa9e18b84f5e2ef