vp9-svc: Allow for nonzero motion on spatial reference.
For nonrd pickmode on a given spatial layer:
if the downsampling filter used for generating the lower spatial layer
is an averaging filter (i.e., nonzero phase), allow for subpel motion on
the spatial (golden) reference to compensate for the shift.
This is done by adding, in addition to the existing (0,0) check on the
spatial reference, a nonzero motion to compensate for the spatial
downsampling shift.

Improvemnt for cases where the downsampling is averaging filter.
In the current code this is only done for generating resolutions <= QVGA.

Avergae improvement for avgPSNR/SSIM on RTC set for speed 7: ~1%.
Gain is larger (~2%) for VGA clips with 2 spatial layers.

Change-Id: I9ec4fa20a38947934fc650594596c25280c3b289