Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> Is there more to be done here? I don't see anything using the new call meaningfully.

GetFrameSize() is called on encoding every frame and it calls the new API to get the encoded frames and push it into the list. After the encoding, the test compares the frame sizes from decoder to those we got in the encoding side (pushed into the list) to see if the API works as intended.

Patch-set: 7
Reviewer: Gerrit User 1171444 <1171444@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>