Bug fix for second reference stats.
Immediately following a key frame the trailing second reference
error in the first pass stats will be based on a reference frame from
the prior key frame group and will thus usually be much larger.

This fix eliminates that effect (which typically triggers a short arf
group immediately after a key frame). It also changes the accounting
for the first frame in each new arf group.

This change gives large gains on a couple of clips that contain mid
sequence key frames (e.g. 6% on 1080P tennis). Overall all sets the
results give a net average gain in PSNR and PSNR-HVS of 0.2-0.5%,
mixed results (+/- 0.15%) for SSIM and are about 0.1-0.7% worse for
Fast SSIM.

Change-Id: I8e00538ac2c0b5c2e7e637903cac329ce5c2a375