vp9: Update to SVC datarate tests.
With recent fixes to rate control for SVC, the
commenting out of buffer underrun can be removed.
This also removes the TODO.

Also replace some of these SVC tests with the corresponding
VGA clip, which has > 400 frames.
For the HD clip: it has only 60 frames but the test was running
up to 300 frames.. Fixed it to 60 frames.

Change-Id: I0a2356a908e8b2271c7a422eb8b15c0d56eec968