vp9-svc: Allow for skip encoding of enhancement layers on the fly.
If an enhancement layer (spatial_layer > 0) has 0 bandwidth,
skip/drop the encoding of the layer. This allows the application
to dynamically disable higher layers for SVC.

Add flag to signal the skip encoding, this is needed to pack the
superframe when the top layer has size 0 (i.e. it was skipped/dropped).

Added datarate unittest.

Change-Id: Idaccdb4aca25ba1d822ed1b4219f94e2e8640d43