Update patch set 1
Patch Set 1:

I did consider this but even on totally static scene the reported error / complexity score is never 0 because the complexity / error score for each mb includes a component allowing for the cost of mode and motion vector etc. So the case you mention of a change to a very small area should not trigger.

I could do some tests to see if I could duplicate the sort of false positive scenario you mention. But... even in the sample clip transitions between the smaller image slides would not be detected if you raise the ratio too much. I think 10:1 would in this case "just" have caught all the key frames but any higher and it would have failed on some of the smaller pictures. So a change to only a few macro blocks almost certainly wont trigger.

I will send you over the first pass stats file from the test sample on monday so you can see what I mean. I wont say a false positive on this is impossible but against that you have the current situation where there are lots of slide shows where we are almost certain to miss key frames and the impact of this is very deleterious.

Patch-set: 1
Tag: mailMessageId=<CALRks=QAOY-i+M=B_daxCVV-8n7Y1-mXbxHpGMtw525p=_Z1mg@mail.gmail.com>