Update patch set 1
Patch Set 1:

Actually nother situation which might trigger and might be more likely than the one you mention could be if a video is up sampled to a higher frame with exact frame duplication.

So say a 10 fps video turned into 30 fps with each frame exactly duplicated 3 times. When is a slide show not a slide show? Normally you might expect a real slide show to show each image for a decent amount of time but the sample he sent some of the images were only up for 1/2 a seconds or so. pretty painful to watch actually.

I will run tests on this on Monday to see if I can create a clip that triggers false positives and also look at what I can do re the worst case outcome we happened to get in his cvbr encode of that clip.

Patch-set: 1
Tag: mailMessageId=<CALRks=Rj7ZKuvNqAymevJcctvNHMqaEqJ_FwtqJsWKwt-4q+Sw@mail.gmail.com>