Update patch set 1
Patch Set 1:

So you are saying a large portion of the blocks are essentially easy to code either intra or inter.

Fine if the images are set on a plain background but then we will be unable to detect key frames if some nice person decides that instead of framing the smaller images against a black background they use a patterned background.

Patch-set: 1
Tag: mailMessageId=<CALRks=TXGiiVcveHDc=DtteXpO5haQP0fQ_LrH8TdJGhzycGXA@mail.gmail.com>