Update patch set 1
Patch Set 1:

I'm saying that we are getting a big benefit from using prior frames as a reference?

If there is a complex pattern in the back we'll get a big benefit in not making them a keyframe,  right?

Patch-set: 1
Tag: mailMessageId=<CANzSYSfH_6NEt6fd_G+rZ_TJsnOGmaoNPCBq+bQSiwSAJcy+EQ@mail.gmail.com>