vp8: Fix to multi-res-encoder for skipping streams.
Add flags to keep track of the disabling/skipping of
streams for the multi-res-encoder. And if on a given stream
the lower spatial stream is skipped, disable the motion vector
reuse for that stream.

Also remove the condition of forcing same frame type across
all streams.

This fix allows skipping/disabling of any stream
(i.e., including base or middle streamr).

Change-Id: Idfa94b32b6d2256932f6602cde19579b8e50a8bd