Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I tried this patch locally and it fixes the bug.
> 
> Though I still don't understand how this could happen...

It would be good to figure out and add a test that covers the case. If there's a particular source clip we could use that. If this should be impossible, though then there may be a deeper issue.

Patch-set: 1
Reviewer: Gerrit User 1002219 <1002219@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>