vp9-svc: Frame dropper for SVC.
If a given spatial layer decides to drop, due to the
buffer/overshoot conditions for that layer, then drop
that current spatial layer and all spatial layers above.

Add last spatial layer encoded to the pkt.
This is useful for RTC applications that enable
frame dropping for SVC.

Enabled frame dropper on all SVC datarate tests.
Change-Id: If828c193f3cb6b1839803fd52fe9fbbda5b5a039