vp9-svc: Improve frame dropper for spatial layers.
SVC frame dropper: modify the logic to allow for individual
spatial layers to drop. This removes the constraint that all
upper spatial layers must drop when a given spatial layer drops.

Disable the inter-layer (spatial) prediction for a given spatial
layer if the previous spatial layer is dropped. And add a flag to
the pkt to indicate whether a spatial layer is encoded or dropped.
This is needed for applications that enable this feature.

Also add the constraint to never drop a spatial layer if its
base layer is a key frame.

Updates to sample encoder (vp9_spatial_svc_encoder) and the
SVC datarate unittests to properly handle frame dropping.

Change-Id: I7d14ccf67b8d014a7abfce5ba3989fc623e94067